Major updates to subsystems, minor package updates. 1st: General compressor subsystem replaces the awkward stanzas in the makefile. 2nd: Compressor information put into config/all.conf 3rd: prepare_initramfs.bash script added to each of the debian based OSes 4th: removed older NVIDIA drivers, put in newer driver 5th: updated zfs to 0.8.4 6th: added in lsint.pl and spark to provide a visualization on interrupt usage
These changes reduced the makefile complexity, and lines of code substantially,
while adding a very flexible capability for compressor selection.

In testing, zstd was found to be significantly faster, in a single threaded manner
than pbzip2, and pigz on all cores.  Hence zstd is now the default, and you can
change it if you wish in the config/all.conf file, or using the COMP=compressor